+++ scripts/deploy.sh
@@ -28,35 +28,78 @@
 #//               0.0.2b1 -- beta-format
 #//               0.0.2a1 -- alpha-format
 
-lib_sh() {
-	filename="$1"
-	"./scripts/lib/$filename.sh" "$2"
-}
+# How to use
+## ./deploy.sh [alpha|beta|patch|minor|major] [--yes <title>]
 
+################################
+## Input from user            ##
+################################
+semver="$1"
+yes=false
+[[ $2 == "--yes" ]] && yes=true
+
+shift 2
+title="$*"
+
+################################
+## Summary result             ##
+################################
+echo "Expected version: $semver"
+echo "Is always yes: $yes"
+echo "What is the title"
+
+################################
+## Helper method              ##
+################################
 lib() {
 	filename="$1"
 	node "./scripts/lib/$filename.js" "$2"
 }
 
-expected="$(lib_sh setVersion)"
-echo "$expected"
-yarn version --new-version "$expected" --no-git-tag-version
+update_version() {
+	local ver="$1"
+	[[ $ver == "alpha" ]] && yarn --silent version:alpha "$version" && return
+	[[ $ver == "beta" ]] && yarn --silent version:beta "$version" && return
+	[[ $ver == "patch" ]] && yarn --silent version:patch "$version" && return
+	[[ $ver == "minor" ]] && yarn --silent version:minor "$version" && return
+	[[ $ver == "major" ]] && yarn --silent version:major "$version" && return
 
+	echo "invalid version" && exit 5
+}
+
+################################
+## Get current version        ##
+################################
+export version
 version="v$(lib "getVersion")"
 
-if lib "promptYN" "create release of version $version"; then
+################################
+## Update new version         ##
+################################
+if test -z "$semver"; then
+	printf "Update version: (alpha|beta|patch|minor|major) "
+	read -r semver
+fi
+
+expected="$(update_version "$semver")"
+yarn version --new-version "$expected" --no-git-tag-version
+
+################################
+## Start deployment           ##
+################################
+if $yes || lib "promptYN" "create release of version $version"; then
 	# check git status
 	status="$(git status --short)"
 
 	if test -z "$status"; then
-		if ! lib "promptYN" "git status must be empty"; then
+		if $yes || (! lib "promptYN" "git status must be empty"); then
 			exit 1
 		fi
 	fi
 
 	# create changelog
 	# bypass gitgo since it not support --next-tag option
-	git-chglog --config ./.gitgo/chglog/config.yml --next-tag "$version" -o "./CHANGELOG.md"
+	git-chglog --config ./.gitgo/chglog/config.yml --next-tag "$expected" -o "./CHANGELOG.md"
 
 	# create jsdoc
 	yarn docs
@@ -65,32 +108,34 @@ if lib "promptYN" "create release of version $version"; then
 
 	# save work
 	git add .
-	git commit -m "[release] Version: $version"
+	git commit -m "[release] Version: $expected"
 
 	# create bin file
 	yarn deploy
 
 	# create git tag
-	git tag "$version"
+	git tag "$expected"
 
 	# update work
-	if lib "promptYN" "execute git push code and tag"; then
+	if $yes || lib "promptYN" "execute git push code and tag"; then
 		git push && git push --tag
 	fi
 
 	# create release
-	if lib "promptYN" "create release in github"; then
+	if $yes || lib "promptYN" "create release in github"; then
 		prerelease=""
-		if lib "promptYN" "This is prerelease"; then
+		if [[ "$semver" == "alpha" ]] || [[ "$semver" == "beta" ]]; then
 			prerelease="--prerelease"
 		fi
-		printf "your release title message is "
-		read -r title
-		message="$(git-chglog --config ./.gitgo/chglog/config.yml "$version")"
+		if test -z "$title"; then
+			printf "your release title message is "
+			read -r title
+		fi
+		message="$(git-chglog --config ./.gitgo/chglog/config.yml "$expected")"
 
 		final="$(printf '%s\n%s' "$title" "$message")"
 
-		hub release create "$prerelease" --message="$final" "$version" -a "./bin/nd-linux" -a "./bin/nd-macos" -a "./bin/nd-win.exe"
+		hub release create "$prerelease" --message="$final" "$expected" -a "./bin/nd-linux" -a "./bin/nd-macos" -a "./bin/nd-win.exe"
 	fi
 else
 	echo "exit"