+++ src/helpers/action.ts
@@ -5,9 +5,8 @@
 
 import { log } from "winston";
 import { PARAM_WRONG_ERR } from "../constants/error.const";
-import { Exception } from "../models/Exception";
+import Throwable, { Exception } from "../models/Exception";
 import { WrapTMC, WrapTM } from "../models/LoggerWrapper";
-import { os } from "pjson";
 
 /**
  * Helper for {@link IfValidate} function
@@ -55,14 +54,24 @@ export const ValidList = (
  *
  * @param e exception to be throw if exist
  */
-export const ThrowIf = (e?: Error) => {
+export const ThrowIf = (e?: Throwable) => {
   if (e) {
-    if (e instanceof Exception) e.printAndExit();
-    else {
-      log(WrapTM("error", "Error", e.message));
+    if (e.printAndExit) {
+      e.printAndExit();
+    } else {
+      log(WrapTM("error", "Error", e.stack ? e.stack : e.message));
       process.exit(1);
     }
   }
+};
+
+export const SaveIf = (e?: Exception) => {
+  if (e && e.save) e.save();
+};
+
+export const Throw = (e: Throwable, message?: string) => {
+  if (message) e.loadString(message);
+  ThrowIf(e);
 };
 
 /**