+++ test/test.js
@@ -10,3 +10,7 @@ exports.setProperty = (object, property, value) => {
   Object.defineProperty(object, property, { value });
   return originalProperty;
 };
+
+exports.TEST_TOKEN =
+  "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoidGVzdHVzZXIiLCJpYXQiOjE1NDAyMTk4MzgsIm5iZiI6MTU0MDIxOTgzOCwiZXhwIjoxNTcxNzc3NDM4LCJpc3MiOiJORC1KUyBtYXN0ZXIiLCJzdWIiOiJORC1KUyIsImp0aSI6IktDLVNLRSJ9.sezZTWbEUXBf3TwrpNl6I3EFow68TppazVOQoiEcJ88";
+exports.TEST_NAME = "admin user admin@nd.com";