+++ CHANGELOG.md
@@ -2,6 +2,73 @@
 ## [Unreleased]
 
 
+<a name="1.0.0-beta.2"></a>
+## [1.0.0-beta.2] - 2018-10-23
+### 1) Features
+- Add validator command and more
+- Add new output to download and fix report error
+- New output now support in raw download
+- Add --raw and --file to init, to setting token and username
+- Add token generator and security update
+- Add Token verify code
+- Implement JWT in app token
+- Add --raw to config path
+- Add warning and error on the end of script too
+- Add resource file JSON format
+- Add resource model
+
+### 2) Improving User Experience
+- Remove verbose log, because it log in file
+- Change default level to verbose
+- Change output color
+- Update logger, no debug + verbose in console
+- Update UI color and add --with-chapter to fetch
+- Add clickable link instead of long link
+- Add log for sold, closed and unknown chapter
+- Add --with-chapter to log result the chapter
+- Add exception summary to raw-download
+- Add short story in novel v1
+- Add OG to html file
+- Change title if chapter name not exist
+
+### 3) Fixes Bug
+- Fix deployment script
+- Fix duplicate version in LOC
+- Fix changelog to next tag
+- Fix error when deploy new version
+- Fix the way to release new version
+- username has space in CI
+- Fix all error test
+- Fix checking type error
+- Fix typo of username output
+- Fix verify error, not jwt id
+- Fix args not found
+- Typo
+- Fix chapter list only completed chapter
+- Fix the way how to check the exception
+- Fix novel not normalize as expected
+- novel name should be back
+- Fix config not found
+
+### 4) Documentation
+- Add LOC for each version
+
+### 5) Code Refactoring
+- Refactor code
+- Move command from action folder
+- Remove log in console
+- Update exception to avoid error on normal throw
+- Call method as promise chain
+- Clean HTML code
+- Make downloader as sequence by promise
+- refactor download novel api and helper
+- Refactor save method
+
+### 6) Libraries
+- Add new command helper
+- Add Inquirer, prompt and fix ^ in package
+
+
 <a name="v1.0.0-beta.1"></a>
 ## [v1.0.0-beta.1] - 2018-10-04
 ### 1) Features
@@ -106,7 +173,8 @@
 - Add gitgo command
 
 
-[Unreleased]: https://github.com/kamontat/nd-js/compare/v1.0.0-beta.1...HEAD
+[Unreleased]: https://github.com/kamontat/nd-js/compare/1.0.0-beta.2...HEAD
+[1.0.0-beta.2]: https://github.com/kamontat/nd-js/compare/v1.0.0-beta.1...1.0.0-beta.2
 [v1.0.0-beta.1]: https://github.com/kamontat/nd-js/compare/v1.0.0-beta.0...v1.0.0-beta.1
 [v1.0.0-beta.0]: https://github.com/kamontat/nd-js/compare/v1.0.0-alpha.1...v1.0.0-beta.0
 [v1.0.0-alpha.1]: https://github.com/kamontat/nd-js/compare/1.0.0-alpha.0...v1.0.0-alpha.1